Added some code that can be ignored.
Code adds a TrajectoryBuilder which is a Builder interface for PathPlanner's implementation of on-the-fly path generation which will be used for making paths while the robot is in teleoperation. This interface introduces the ability to use method chaining for easier creation of new paths that are more readable. Moreover, added an autonomous abstract class that will act as an interface for autonomous code used in commands. Right now it has the various important position that the robot could be at on the field.